Validations updated

+++ src/Controller/Api/UsersController.php
@@ -4,8 +4,9 @@ namespace App\Controller\Api;
 use App\Controller\Api\ApiController;
 use Cake\Network\Exception\BadRequestException;
 use Cake\Network\Exception\MethodNotAllowedException;
-use Cake\Network\Exception\Exception;
+use Cake\Core\Exception\Exception;
 use Cake\Network\Exception\NotFoundException;
+use Cake\Network\Exception\UnauthorizedException;
 use Cake\Auth\DefaultPasswordHasher;
 use Firebase\JWT\JWT;
 use Cake\Utility\Security;
@@ -77,6 +78,10 @@ class UsersController extends ApiController
         $user = $this->Users->newEntity();
         $data = $this->request->getData();
 
+        if(!isset($data['role_id'])){
+            throw new Exception("No role provided for the user.");
+        }
+
         if($data['role_id'] == 3){
             $data['experts'] = [[]];
             $user = $this->Users->patchEntity($user, $data, ['associated' => 'Experts']);
@@ -128,6 +133,10 @@ class UsersController extends ApiController
      */
     public function delete($id = null)
     {
+        if($this->Auth->user('role_id') != 1){
+            throw new UnauthorizedException("You're Not alloweed to access this method.");
+        }
+
         $this->request->allowMethod(['post', 'delete']);
         $user = $this->Users->get($id);
         if ($this->Users->delete($user)) {
@@ -157,7 +166,7 @@ class UsersController extends ApiController
         }
 
         $this->loadModel('Users');
-        $user = $this->Users->find()->where(['email' => $email])->first();
+        $user = $this->Users->find()->where(['email' => $email])->contain(['Experts'])->first();
         $token=null;
         $success=false;
 
@@ -172,6 +181,8 @@ class UsersController extends ApiController
         }else{
             throw new NotFoundException(__('ENTITY_DOES_NOT_EXISTS','User'));
         }
+
+        $this->request->session()->write('User',$user->toArray()); 
 
         $this->set([
            'success' => $success,
